refactor: FinishPage 데이터 페칭 로직 훅으로 분리
- submissionResult 조회 useEffect 로직을 useSubmissionResult 훅으로 추출
